extract single components from Schedule
