+++ client/styles/main.scss
@@ -43,6 +43,9 @@
 @import '../../shared/components/SchedulePage/index';
 @import '../../shared/components/SchedulePageOverview/index';
 @import '../../shared/components/Schedule/index';
+@import '../../shared/components/ScheduleItem/index';
+@import '../../shared/components/ScheduleDetail/index';
+@import '../../shared/components/CommunityEvents/index';
 @import '../../shared/components/SpeakerPageList/index';
 @import '../../shared/components/SpeakerPageSpeaker/index';
 @import '../../shared/components/SpeakerPagePlaceholder/index';
