+++ shared/components/ScheduleItem/index.js
@@ -1,0 +1,35 @@
+import React, { PropTypes } from 'react';
+
+import ScheduleDetail from '../ScheduleDetail';
+
+const ScheduleItem = ({ item }) => (
+  <div className="ScheduleItem">
+    <div className="ScheduleItem__header">
+      <time className="ScheduleItem__header__time">
+        {item.time}
+      </time>
+      <h3 className="ScheduleItem__header__title">
+        {item.title}
+      </h3>
+    </div>
+    <div className="ScheduleItem__details">
+      <p className="ScheduleItem__details__description">
+        {item.description}
+      </p>
+      <ul className="ScheduleItem__details__list">
+        {item.details.map((detail, index) => <ScheduleDetail key={index} detail={detail} />)}
+      </ul>
+    </div>
+  </div>
+);
+
+ScheduleItem.propTypes = {
+  item: PropTypes.shape({
+    time: PropTypes.string,
+    title: PropTypes.string,
+    description: PropTypes.description,
+    details: PropTypes.arrayOf(PropTypes.shape(ScheduleDetail.propTypes)),
+  }),
+};
+
+export default ScheduleItem;
